Replacing StackoverflowPage.java and StackoverflowSearchResultsTest.java by TutorialspointPage.java and TutorialspointSearchResultsTest.java.
