+++ src/test/resources/testng.xml
@@ -12,7 +12,7 @@
             </class>
             <class name="tests.WikipediaImageTest">
             </class>
-            <class name="tests.StackoverflowSearchResultsTest">
+            <class name="tests.TutorialspointSearchResultsTest">
             </class>
         </classes>
     </test>
@@ -25,7 +25,7 @@
             </class>
             <class name="tests.WikipediaImageTest">
             </class>
-            <class name="tests.StackoverflowSearchResultsTest">
+            <class name="tests.TutorialspointSearchResultsTest">
             </class>
         </classes>
     </test>
@@ -38,7 +38,7 @@
             </class>
             <class name="tests.WikipediaImageTest">
             </class>
-            <class name="tests.StackoverflowSearchResultsTest">
+            <class name="tests.TutorialspointSearchResultsTest">
             </class>
         </classes>
     </test>
